The writing of wal log should never wait for logical walsender flush.
Here explains why we need this modfication for file syncrep.c:
In each walsender process, there is one structure pointer MyWalSnd which
points to the structure of WalSnd, and if the walsender is a phsical
walsender, then the MyWalSnd->is_for_gp_walreceiver is set to true.
In function SyncRepWaitForLSN(), if the process is forked from coordinator
postgres, then it tries to wait for candidate whose walsnd->is_for_gp_walreceiver
is set to true which means it only wants to wait for the lsn to be confirmed by
the physical walreceiver, not including logical walreceiver.
But in coordinator, in logical walsender process, the value of
MyWalSnd->sync_standby_priority is 0, which means not listed. And in function
SyncRepGetSyncRecPtr()->SyncRepGetCandidateStandbys(), it will search all of
the standby candidates, and if logical walsender process is selected, that
means the updates of value of WalSndCtl->lsn[1] will be effected by logical
walsender, and for coordinator, the update of lsn is slow, so it will cause
the updating of the shared memory value WalSndCtl->lsn[1] slowly, then in
function SyncRepWaitForLSN(), it will wait for the slowest updating lsn to be
updated which is updated in logical walsender which is not what we expect.
So we need to add one another condition for searching the candicate standby
in function SyncRepGetCandidateStandbys().

The writing of wal log should never wait for logical walsender flush.

If logical walsender process is selected, that means the updates
of value of WalSndCtl->lsn[1] will be effected by logical walsender
which is not expected. Because the update of the logical walsender's
structure value of walsnd->flush is much slower than physical.
This may cause no matter coordinator or segment hung in function
SyncRepWaitForLSN(), because the MyProc->waitLSN may be one value that
is larger than logical walsender's flush, but smaller than physical
walsender's flush. So it will hangs on waiting for logical walsender
to catch up, however, logical does not replay all of the wal log, so
it will hang forever.

If without this check, with a logical replication started in each segment
and coordinator:
Coordinator: normal commit will hang.
For segment: if one of the segment mirror is killed, and the gproverseg
will cause all of the walsender restart, then cause segment hang
after commit new changes.